requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
httpx>=0.25.0
python-multipart>=0.0.6
beautifulsoup4>=4.12.0
lxml>=5.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0
aiofiles>=23.2.0
Pillow>=10.0.0
google-genai
selenium>=4.15.0
webdriver-manager>=4.0.0
